Factor out Kore.OnePath.StrategyPattern
