+++ kore/src/Kore/OnePath/Step.hs
@@ -10,11 +10,6 @@ Portability : portable
 module Kore.OnePath.Step
     ( -- * Primitive strategies
       Prim (..)
-    , StrategyPattern (..)
-    , StrategyPatternTransformer (..)
-    , CommonStrategyPattern
-    , extractUnproven
-    , extractStuck
     , simplify
     , transitionRule
     , onePathFirstStep
@@ -26,10 +21,8 @@ import           Control.Applicative
                  ( Alternative (..) )
 import qualified Control.Monad.Trans as Monad.Trans
 import qualified Data.Foldable as Foldable
-import           Data.Hashable
 import qualified Data.Set as Set
 import qualified Data.Text.Prettyprint.Doc as Pretty
-import           GHC.Generics
 
 import           Kore.Attribute.Symbol
                  ( StepperAttributes )
@@ -42,6 +35,7 @@ import           Kore.Internal.Pattern
                  ( Pattern )
 import qualified Kore.Internal.Pattern as Pattern
 import           Kore.Internal.TermLike
+import           Kore.OnePath.StrategyPattern
 import           Kore.Predicate.Predicate
                  ( Predicate )
 import qualified Kore.Predicate.Predicate as Predicate
@@ -93,87 +87,6 @@ debugString :: (Show patt, Show rewrite) => Prim patt rewrite -> String
 debugString Simplify = "Simplify"
 debugString s@(RemoveDestination _) = show s
 debugString (ApplyWithRemainders _) = "ApplyWithRemainders"
-
-{- | A pattern on which a rule can be applied or on which a rule was applied.
-
-As an example, when rewriting
-
-@
-if x then phi else psi
-@
-
-with these rules
-
-@
-if true  then u else v => u
-if false then u else v => v
-@
-
-there would be two 'RewritePattern's, @phi and x=true@ and @psi and x=false@.
-If only the first rule was present, the results would be a 'RewritePattern' with
-@phi and x=true@ and a 'Remainder' with @psi and not x=true@.
-
-When rewriting the same pattern with an rule that does not match, e.g.
-
-@
-x + y => x +Int y
-@
-
-then rewriting produces no children.
-
--}
-data StrategyPattern patt
-    = RewritePattern !patt
-    -- ^ Pattern on which a normal 'Rewrite' can be applied. Also used
-    -- for the start patterns.
-    | Stuck !patt
-    -- ^ Pattern which can't be rewritten anymore.
-    | Bottom
-    -- ^ special representation for a bottom rewriting/simplification result.
-    -- This is needed when bottom results are expected and we want to
-    -- differentiate between them and stuck results.
-  deriving (Show, Eq, Ord, Generic)
-
-{- | Extract the unproven part of a 'StrategyPattern'.
-
-Returns 'Nothing' if there is no remaining unproven part.
-
- -}
-extractUnproven :: StrategyPattern patt -> Maybe patt
-extractUnproven (RewritePattern p) = Just p
-extractUnproven (Stuck          p) = Just p
-extractUnproven Bottom             = Nothing
-
--- | Extract the 'Stuck' part of a 'StrategyPattern'.
-extractStuck :: StrategyPattern patt -> Maybe patt
-extractStuck (Stuck p) = Just p
-extractStuck _         = Nothing
-
-data StrategyPatternTransformer patt a =
-    StrategyPatternTransformer
-        { rewriteTransformer :: patt -> a
-        , stuckTransformer :: patt -> a
-        , bottomValue :: a
-        }
-
--- | Catamorphism for 'StrategyPattern'
-strategyPattern
-    :: StrategyPatternTransformer patt a
-    -> StrategyPattern patt
-    -> a
-strategyPattern
-    StrategyPatternTransformer
-        {rewriteTransformer, stuckTransformer, bottomValue}
-  =
-    \case
-        RewritePattern patt -> rewriteTransformer patt
-        Stuck patt -> stuckTransformer patt
-        Bottom -> bottomValue
-
--- | A 'StrategyPattern' instantiated to 'Pattern Variable' for convenience.
-type CommonStrategyPattern = StrategyPattern (Pattern Variable)
-
-instance Hashable patt => Hashable (StrategyPattern patt)
 
 -- | Apply the rewrites in order. The first one is applied on the start pattern,
 -- then each subsequent one is applied on the remainder of the previous one.
@@ -396,8 +309,7 @@ onePathFirstStep destination rewrites =
     Strategy.sequence
         [ Strategy.apply simplify
         , Strategy.apply (removeDestination destination)
-        , Strategy.apply
-            (applyWithRemainder rewrites)
+        , Strategy.apply (applyWithRemainder rewrites)
         , Strategy.apply simplify
         ]
